+++ dask_geomodeling/raster/elemwise.py
@@ -58,8 +58,9 @@ class BaseElementwise(RasterBlock):
         if start is not None and stop is not None:
             # limit request to self.period so that resulting data is aligned
             period = self.period
-            request["start"] = max(start, period[0])
-            request["stop"] = min(stop, period[1])
+            if period is not None:
+                request["start"] = max(start, period[0])
+                request["stop"] = min(stop, period[1])
 
         process_kwargs = {"dtype": self.dtype.name, "fillvalue": self.fillvalue}
         sources_and_requests = [(source, request) for source in self.args]