+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of dask-geomodeling
 2.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix empty response of TemporalAggregate and Cumulative.
 
 
 2.1.0 (2019-11-15)